Added contracts compilation test
