+++ .editorconfig
@@ -8,6 +8,9 @@ indent_size = 4
 insert_final_newline = true
 trim_trailing_whitespace = true
 
+[*.json]
+indent_size = 2
+
 [*.md]
 max_line_length = off
 trim_trailing_whitespace = false
