+++ .eslintrc.yml
@@ -2,6 +2,7 @@ env:
   node: true
   amd: true
   es6: true
+  mocha: true
 extends: 'eslint:recommended'
 parser: 'babel-eslint'
 parserOptions:
