+++ contracts/empty.sol
@@ -1,0 +1,4 @@
+pragma solidity ^0.5.7;
+
+contract Empty {
+}
