+++ contracts/external-lib.sol
@@ -1,0 +1,7 @@
+pragma solidity ^0.5.7;
+
+library External {
+  function increment(uint256 _n) public pure returns (uint256) {
+    return (_n + 1);
+  }
+}
