+++ contracts/import.sol
@@ -1,0 +1,5 @@
+pragma solidity ^0.5.7;
+
+import "./imported.sol";
+
+contract Import is Imported {}
