+++ contracts/imported.sol
@@ -1,0 +1,3 @@
+contract Imported {
+    function importedFunction() public returns (bool) {return true;}
+}
