+++ contracts/no-pragma.sol
@@ -1,0 +1,6 @@
+contract NoPragma {
+    function f() public {
+        selfdestruct(msg.sender);
+    }
+}
+
