+++ lib/report.js
@@ -1,10 +1,5 @@
 const chalk = require('chalk');
 const eslintHelpers = require('./eslint');
-const fs = require('fs');
-const https = require('https');
-const parser = require('solidity-parser-antlr');
-const path = require('path');
-const releases = require('./releases');
 const SourceMappingDecoder = require('remix-lib/src/sourceMappingDecoder');
 
 const mythx2Severity = {
@@ -185,7 +180,7 @@ const convertMythXReport2EsIssue = (report, data) => {
     return Object.values(results);
 };
 
-const doReport = (data, issues) => {
+const formatIssues = (data, issues) => {
     const eslintIssues = issues
         .map(report => convertMythXReport2EsIssue(report, data))
         .reduce((acc, curr) => acc.concat(curr), []);
@@ -200,129 +195,6 @@ const doReport = (data, issues) => {
     }
 };
 
-const getSolidityVersion = fileContents => {
-    try {
-        const ast = parser.parse(fileContents, {});
-        let solidityVersion = releases.latest;
-        let reg = RegExp(/[><=^]/, 'g');
-
-        for (let n of ast.children) {
-            if ((n.name === 'solidity') && (n.type === 'PragmaDirective')) {
-                solidityVersion = n.value;
-                if (!reg.test(solidityVersion)) {
-                    return solidityVersion;
-                }
-                break;
-            }
-        }
-
-        if (solidityVersion !== releases.latest) {
-            solidityVersion = solidityVersion.replace(/[\^v]/g, '');
-            let upperLimit = 'latest';
-
-            if (solidityVersion.indexOf('<') !== -1) {
-                if (solidityVersion.indexOf('<=') !== -1) {
-                    solidityVersion = solidityVersion.substring(solidityVersion.length - 5, solidityVersion.length);
-                } else {
-                    upperLimit = solidityVersion.substring(solidityVersion.length - 5, solidityVersion.length);
-                }
-            } else if (solidityVersion.indexOf('>') !== -1) {
-                solidityVersion = releases.latest;
-            } else {
-                upperLimit = '0.' + (parseInt(solidityVersion[2]) + 1).toString() + '.0';
-            }
-
-            if (upperLimit !== 'latest') {
-                if (upperLimit === '0.6.0' || upperLimit === '0.7.0') {
-                    solidityVersion = releases.latest;
-                } else if (upperLimit === '0.5.0') {
-                    solidityVersion = '0.4.25';
-                } else if (upperLimit === '0.4.0') {
-                    solidityVersion = '0.3.6';
-                } else if (upperLimit === '0.3.0') {
-                    solidityVersion = '0.2.2';
-                } else {
-                    let x = parseInt(upperLimit[upperLimit.length - 1], 10) - 1;
-                    solidityVersion = '';
-                    for (let i = 0; i < upperLimit.length - 1; i++) {
-                        solidityVersion += upperLimit[i];
-                    }
-                    solidityVersion += x.toString();
-                }
-            }
-        }
-
-        return solidityVersion;
-    } catch (error) {
-        for (let err of error.errors) {
-            console.error(err.message);
-        }
-    }
-};
-
-const getImportPaths = source => {
-    let matches = [];
-    let ir = /^(.*import){1}(.+){0,1}\s['"](.+)['"];/gm;
-    let match = null;
-
-    while ((match = ir.exec(source))) {
-        matches.push(match[3]);
-    }
-
-    return matches;
-};
-
-const removeRelativePathFromUrl = url => url.replace(/^.+\.\//, '').replace('./', '');
-
-/* Dynamic linking is not supported. */
-
-const regex = new RegExp(/__\$\w+\$__/,'g');
-const address = '0000000000000000000000000000000000000000';
-const replaceLinkedLibs = byteCode => byteCode.replace(regex, address);
-
-/* Get solc-js of specific version from temp directory, if exists */
-
-const loadSolcVersion = (version, callback) => {
-    const tempDir = path.dirname(require.main.filename) + '/.temp/';
-    const solcPath =  tempDir + version + '.js';
-
-    if (fs.existsSync(solcPath)) {
-        callback(fs.readFileSync(solcPath).toString());
-    } else {
-        /* Create `.temp` directory if it doesn't exist */
-
-        if (!fs.existsSync(tempDir)) {
-            fs.mkdirSync(tempDir);
-        }
-
-        /* Get the solc remote version snapshot of the specified version */
-
-        const solcJs = fs.createWriteStream(solcPath);
-        const url = 'https://ethereum.github.io/solc-bin/bin/soljson-' + version + '.js';
-
-        https
-            .get(url, (response) => {
-                if (response.statusCode !== 200) {
-                    throw new Error('Error retrieving binary: ' + response.statusMessage);
-                } else {
-                    const stream = response.pipe(solcJs);
-
-                    stream.on('finish', () => {
-                        callback(fs.readFileSync(solcPath).toString());
-                    });
-                }
-            })
-            .on('error',  (error) => {
-                throw new Error('Error fetching binary: ' + error);
-            });
-    }
-};
-
 module.exports = {
-    doReport,
-    getImportPaths,
-    getSolidityVersion,
-    loadSolcVersion,
-    removeRelativePathFromUrl,
-    replaceLinkedLibs
+    formatIssues
 };
